requirements.txt
requests==2.28.1
urllib3==1.26.11
python-telegram-bot==13.15
python-dotenv==1.0.0
